+++ src/radiant/source/resolvers/__init__.py
@@ -2,7 +2,6 @@
 
 from radiant.source.resolvers.direct import resolve_direct_radiance
 from radiant.source.resolvers.geometry import resolve_geometry
-from radiant.source.resolvers.intensity import resolve_direct_intensity
 from radiant.source.resolvers.physical import resolve_physical_object
 from radiant.source.resolvers.resolved_target import ResolvedTarget
 from radiant.source.resolvers.sub_pixel import resolve_sub_pixel
@@ -11,7 +10,6 @@ __all__ = [
     "ResolvedTarget",
     "resolve_direct_radiance",
     "resolve_geometry",
-    "resolve_direct_intensity",
     "resolve_physical_object",
     "resolve_sub_pixel",
 ]